requirements.txt
# Core ML & Data Science
xgboost>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# ML Ops
mlflow>=2.9.0
boto3>=1.28.0

# Database
psycopg2-binary>=2.9.0

# Messaging
kafka-python>=2.0.0
paho-mqtt>=1.6.0

# Monitoring
prometheus-client>=0.18.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
